requirements.txt
asgiref==3.4.1
Django==3.2.9
django-extensions==3.1.5
django-import-export==2.6.1
djangorestframework==3.12.4
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
